+++ ts/src/_exports/public-sdk.ts
@@ -326,6 +326,7 @@ export {
   buildPlacePositionConditionalOrder,
   buildPlaceStopLoss,
   buildRegisterTrader,
+  buildOnboardTraderDelegated,
   registerTrader,
   buildSyncParentToChild,
   buildTransferCollateral,